Build admb / muppet on installation
Instead of assuming that muppet will be available on path, build both
ADMB & Muppet_HCR on installation, and use that instead.

Bake in the version of ADMB & Muppet_HCR the package uses, so a version of
rmuppet directly corresponds to a version of Muppet_HCR.

For development ADMB_HOME & MUPPET_HOME can be modified to a local path
so you don't have to re-install admb with every package install.